[np] extend mobile support, prepare for further mobile updates
- added redirects for standardfeeds to support 1.5.20.
- added support for /proxy/ytbt, seen in 1.3 and 1.4 apps.
(#106 - thanks, gohoski!)
! the apps need slight changes to work with yt2009 responses !
! read more in apk_setup.md !

^ both tested on android 2.1, mobilehelper untested;
don't expect it to work.

- fixed basic mobile streaming options. (thanks, ISM_64!)
- html5 sabr will prebuffer more data in advance for videos
that need it. this will reduce random stutters
while watching videos.
- fixed #286 (thanks, smartiphone7!)
- partially cleaned up a few innertube generator protobufs.
- removed protoc dependency for yt2009m.

[!] this update includes a minor change related to
communication with the sync server. update to keep
your sync functionality properly working in the future.

//marker: 1.23.2

+++ back/yt2009utils.js
@@ -1415,6 +1415,7 @@ module.exports = {
                     `-i "${__dirname}/${audioFile}"`,
                     `-c:v copy -c:a copy`,
                     `-map 0:v -map 1:a`,
+                    `-movflags +faststart`,
                     `"${__dirname}/../assets/${fname}.mp4"`
                 ].join(" ")
 
@@ -1686,6 +1687,7 @@ module.exports = {
                     fetchNextPart()
                 }, 500)
             }).then(r => {
+                if(!r) return;
                 if(r.status == 403) {
                     stream.end()
                     try {fs.unlinkSync(out)}catch(error){
